+++ asosnimaux-bdd/src/controllers/article.controllers.js
@@ -2,18 +2,8 @@ import { ArticleDB } from "../databases/article.db.js";
 import { areStringsFilled } from "../utils/string.utils.js";
 import formidable from "formidable";
 import { deleteImg, setImgUrl } from "../utils/formidable.utils.js";
-
-const create_ = async ({ body: { name, location, description, pictureURL, pictureCaption, userID } }, res) => {
-  const areStrings = areStringsFilled([name, location, description, pictureURL, pictureCaption]);
-
-  if (!areStrings) return res.status(403).json({ message: `Missing data` });
-
-  const response = await ArticleDB.create(name, location, description, pictureURL, pictureCaption, userID);
-
-  const { result, error } = response;
-
-  return res.status(error ? 500 : 200).json({ message: error ? error : `New article successfully created`, result });
-}
+import isUUID from "validator/lib/isUUID.js";
+import UUID from "../constants/uuid.const.js";
 
 const create = async (req, res) => {
   const form = formidable({
@@ -23,7 +13,7 @@ const create = async (req, res) => {
     maxFileSize: 5 * 1024 * 1024, // 5MB
     filter: opts => {
       const { mimetype } = opts;
-      return mimetype === "image/png" || mimetype === "image/jpg" || mimetype === "image/jpeg";
+      return mimetype === "image/png" || mimetype === "image/jpg" || mimetype === "image/jpeg" || mimetype === "image/webp";
     }
   });
 
@@ -47,6 +37,8 @@ const create = async (req, res) => {
   const filePath = files.newArticleImg[0].filepath;
   const picture_url = setImgUrl(filePath, "articles");
   const { name, location, description, picture_caption } = fields;
+
+  if (!isUUID(userID, UUID.VERSION)) return res.status(400).json({ error: "Invalid UUID format" });
 
   const areStrings = areStringsFilled([name[0], location[0], description[0], picture_url, picture_caption[0]]);
   if (!areStrings) return res.status(403).json({ message: `Missing data` });
@@ -92,6 +84,8 @@ const read = async (req, res) => {
 }
 
 const readOne = async ({ params: { articleID } }, res) => {
+  if (!isUUID(articleID, UUID.VERSION)) return res.status(400).json({ error: "Invalid UUID format" });
+
   const response = await ArticleDB.readOne(articleID);
 
   const { result, error } = response;
@@ -110,8 +104,9 @@ const readOne = async ({ params: { articleID } }, res) => {
 }
 
 const update = async ({ body: { name, location, description, articleID } }, res) => {
-  const areStrings = areStringsFilled([name, location, description]);
+  if (!isUUID(articleID, UUID.VERSION)) return res.status(400).json({ error: "Invalid UUID format" });
 
+  const areStrings = areStringsFilled([name, location, description]);
   if (!areStrings) return res.status(403).json({ message: `Missing data` });
 
   const response = await ArticleDB.update(name, location, description, articleID);
@@ -127,11 +122,13 @@ const update = async ({ body: { name, location, description, articleID } }, res)
 }
 
 const deleteOne = async ({ params: { articleID } }, res) => {
+  if (!isUUID(articleID, UUID.VERSION)) return res.status(400).json({ error: "Invalid UUID format" });
+
   const response = await ArticleDB.deleteOne(articleID);
 
   const { imgPathResult, error } = response;
   // Opti with -> transaction (SQL)
-  const err = await deleteImg(imgPathResult);
+  const err = await deleteImg(imgPathResult[0].picture_url);
   if (err) return res.status(403).json({ message: err });
 
   return res.status(error ? 500 : 200).json({ message: error ? error : `Article with id ${articleID} deleted successfully` });
